requirements.txt
requests
yaml8
prometheus_client
yo-fluq
pydantic